- Make settings more robust; -minor fixes; prepare FlashROMode

+++ timecircuits-A10001986/speeddisplay.cpp
@@ -564,7 +564,7 @@ void speedDisplay::setSpeed(int8_t speedNum)
 }
 
 #ifdef TC_HAVETEMP
-void speedDisplay::setTemperature(double temp)
+void speedDisplay::setTemperature(float temp)
 {
     char buf[8];
     char alignBuf[20];
@@ -577,7 +577,7 @@ void speedDisplay::setTemperature(double temp)
         else if(temp <= -10.0) setText("Lo");
         else if(t >= 100.0) setText("Hi");
         else if(temp >= 10.0 || temp < 0.0) {
-            t = (int)((double)round(temp));
+            t = (int)((float)round(temp));
             sprintf(buf, "%d", t);
             setText(buf);
         } else {
@@ -590,7 +590,7 @@ void speedDisplay::setTemperature(double temp)
         else if(temp <= -100.0) setText("Low");
         else if(t >= 1000.0) setText("Hi");
         else if(temp >= 100.0 || temp <= -10.0) {
-            t = (int)((double)round(temp));
+            t = (int)((float)round(temp));
             sprintf(buf, "%d", t);
             setText(buf);
         } else {
